+++ ui/src/components/graph/graph.tsx
@@ -22,9 +22,9 @@ const nodeTypes = {
 type Props = {
   data: Elements;
   nodeId: string;
-  isFeatureGraph: boolean;
+  height: number;
 }
-const Graph: React.FC<Props> = ({ data, nodeId, isFeatureGraph }) => {
+const Graph: React.FC<Props> = ({ data, nodeId, height }) => {
   const [, setURLSearchParams] = useSearchParams();
 
   const { layoutedElements, elementMapping } = getLayoutedElements(data);
@@ -65,23 +65,6 @@ const Graph: React.FC<Props> = ({ data, nodeId, isFeatureGraph }) => {
     setElements(values);
   };
 
-  // calculate the height of the graph by finding the maximum y position of the nodes and adding some padding to take the size of the node into account
-  const calculateHeight = () => {
-    if (isFeatureGraph) {
-      var padding = 200;
-      var max = 0;
-      for (let index = 0; index < elements.length; index++) {
-        const element = elements[index];
-        if (isNode(element) && element.position.y > max) {
-          max = element.position.y
-        }
-      }
-      return max + padding;
-    } else {
-      return window.innerHeight - 250;
-    }
-  }
-  
   // Highlight path of selected node, including all linked up and down stream nodes
   const highlightPath = (node: Node, check: boolean): void => {
     const checkElements = check ? layoutedElements : elements;
@@ -164,7 +147,7 @@ const Graph: React.FC<Props> = ({ data, nodeId, isFeatureGraph }) => {
     <div className="lineage-graph">
         <ReactFlowProvider>
           <ReactFlow
-            style={ { height: calculateHeight(), width: "100%" } }
+            style={ { height: height, width: "100%" } }
             elements={ elements }
             snapToGrid
             snapGrid={ [15, 15] }
