+++ ui/src/components/graph/utils.ts
@@ -1,5 +1,6 @@
 import dagre from 'dagre';
 import { ArrowHeadType, Edge, Elements, isNode, Node, Position, } from 'react-flow-renderer';
+import { FeatureLineage } from "../../models/model";
 
 const DEFAULT_WIDTH = 172;
 const DEFAULT_HEIGHT = 36;
@@ -7,6 +8,53 @@ const DEFAULT_HEIGHT = 36;
 type getLayoutElementsRet = {
   layoutedElements: Elements,
   elementMapping: Record<string, number>,
+};
+
+const getElements = (lineageData: FeatureLineage, featureType: string|null) => {
+  if (lineageData.guidEntityMap === null && lineageData.relations === null) {
+    return;
+  }
+
+  const elements: Elements = [];
+  const elementObj: Record<string, string> = {};
+
+  for (let index = 0; index < Object.values(lineageData.guidEntityMap).length; index++) {
+    const currentNode: any = Object.values(lineageData.guidEntityMap)[index];
+
+    if (currentNode.typeName === "feathr_workspace_v1") {
+      continue; // Open issue: should project node get displayed as well?
+    }
+
+    const nodeId = currentNode.guid;
+
+    // If toggled feature type exists, skip other types
+    if (featureType && featureType !== "all_nodes" && currentNode.typeName !== featureType) {
+      continue;
+    }
+
+    const node = generateNode({
+      index,
+      nodeId,
+      currentNode
+    });
+
+    elementObj[nodeId] = index?.toString();
+
+    elements.push(node);
+  }
+
+  for (let index = 0; index < lineageData.relations.length; index++) {
+    var { fromEntityId: from, toEntityId: to, relationshipType } = lineageData.relations[index];
+    if (relationshipType === "Consumes") [from, to] = [to, from];
+    const edge = generateEdge({ obj: elementObj, from, to });
+    if (edge?.source && edge?.target) {
+      if (relationshipType === "Consumes" || relationshipType === "Produces") {
+        elements.push(edge);
+      }
+    }
+  }
+
+  return elements;
 };
 
 const getLayoutedElements = (elements: Elements, direction = 'LR'): getLayoutElementsRet => {
@@ -112,6 +160,7 @@ export {
   generateEdge,
   generateNode,
   getLayoutedElements,
+  getElements
 };
 
 export const findNodeInElement = (nodeId: string | null, elements: Elements): Node | null => {
